+++ apps/api/src/services/platform/platform.module.ts
@@ -1,0 +1,11 @@
+import { PrismaModule } from '@ghostfolio/api/services/prisma.module';
+import { Module } from '@nestjs/common';
+
+import { PlatformService } from './platform.service';
+
+@Module({
+  exports: [PlatformService],
+  imports: [PrismaModule],
+  providers: [PlatformService]
+})
+export class PlatformModule {}