+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved the portfolio evolution chart (ignore first item)
+- Improved the accounts import by handling the platform
+
+### Fixed
+
+- Fixed an issue with more than 50 activities in the activities import (`dryRun`)
 
 ## 1.249.0 - 2023-03-27
 
@@ -245,7 +250,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support to export accounts
-- Added suport to import accounts
+- Added support to import accounts
 
 ### Changed